requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.18
botocore==1.34.18
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
django-redis==5.4.0
djangorestframework==3.14.0
gevent==23.9.1
greenlet==3.0.3
install==1.3.5
jmespath==1.0.1
kombu==5.3.5
prompt-toolkit==3.0.43
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic-bin==0.4.14
pytz==2023.3.post1
redis==5.0.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.1
